Add deliberately failing unit test.
Previous tests written as part of this project were functional tests that were run directly with the functional_tests.py file. However, unit tests are run by the Django-specific automated test runner. We need to ensure the automated test runner is working correctly so that we know that the unit tests we are writing will definitely be run.