Display input field in home page template.
To allow a user to enter a to-do list item, the home page template needs to have an input field. A table is required to display any to-dos that have been successfully added to the to-do list.

Functional tests for the home page have been refactored to add a more descriptive error message to aid debugging.

Future work will process the user input.